Proxy tunnel handler and HTTP request sending
An interface has been implemented that is responsible for managing proxy connections or tunnels, so that it can send HTTP request lifecycle events so that it can be used as an intermediate point between the client connection (tunnel_proxy) and server.

In addition to implementing some error controls during the aforementioned life cycles.

However, much of the progress is unfinished, I know that it is almost time to start testing the functionality of both parts (api_reverse_proxy) and (tunnel_proxy)